[#127]  gitlabmontor job information now in job module
