+++ lib/gitlab_monitor/gitlab_monitor.ex
@@ -239,154 +239,6 @@ defmodule CncfDashboardApi.GitlabMonitor do
   end
 
  @doc """
-  Gets a job names in order of precedence for a specific project based on `project_name`.
-
-  Precendence is from least to most important 
-
-  Returns `["job_name1", "job_name2"]`
-  """
-  def monitored_job_list(project_name) do
-    cloud_list = CncfDashboardApi.YmlReader.GitlabCi.gitlab_pipeline_config()
-    pipeline_config = Enum.find(cloud_list, fn(x) -> x["pipeline_name"] == project_name end) 
-    pipeline_config["status_jobs"]
-  end
-
- @doc """
-  Gets a list of migrated jobs based on based on `pipeline_id` and `job_names`.
-
-  job_names is a list of strings denoting the job name
-
-  A migration from gitlab must have occured before calling this function in order to get 
-  valid jobs 
-
-  Returns `[%job1, %job2]`
-  """
-  def monitored_jobs(job_names, pipeline_id) do
-    Logger.info fn ->
-      "monitored_job_list job_names: #{inspect(job_names)}"
-    end
-    # get all the jobs for the internal pipeline
-    jobs = Repo.all(from pj in CncfDashboardApi.PipelineJobs, 
-                              where: pj.pipeline_id == ^pipeline_id,
-                              where: pj.name in ^job_names)
-    Logger.info fn ->
-      "monitored_job_list monitored_jobs: #{inspect(jobs)}"
-    end
-
-    # sort by job_names
-    job_names
-    |> Enum.reduce([], fn(job_name, acc) ->
-      job = Enum.find(jobs, fn(x) -> x.name =~ job_name end) 
-        if job do
-          [job | acc]
-        else
-          acc
-        end
-    end)
-    |> Enum.reverse
-  end
-
-  def badge_status_by_pipeline_id(monitor_job_list, child_pipeline, _cloud, internal_pipeline_id) do
-
-    Logger.info fn ->
-      "badge_status_by_pipeline_id monitored_job_list: #{inspect(monitor_job_list)}"
-    end
-    %{:status => status, :job => _} = monitored_jobs(monitor_job_list, internal_pipeline_id)
-                  |> status_job(child_pipeline)
-     status
-  end
-
- @doc """
-  Gets the status and job based on precendence based on `monitored_jobs` and `child_pipeline`  
-
-  If any job in the status jobs list has a status of failed, return failed
-  else if any job in the list has a status of running, return running
-  else 
-     if not child pipeline
-       if all jobs are a success, return success
-
-  Returns :nojob if no job found
-
-  Returns `%{:status => "success", :job => %PipelineJob}`
-  """
-  def status_job(monitored_jobs, child_pipeline) do
-    Logger.info fn ->
-      "status_job monitored_jobs: #{inspect(monitored_jobs)}"
-    end
-    # loop through the jobs list in the order of precedence
-    # monitor_job_list e.g. ["e2e", "App-Deploy"]
-    # create status string e.g. "failure"
-    # If any job in the status jobs list has a status of failed, return failed
-    # else if any job in the list has a status of running, return running
-    # else 
-    #    if not child pipeline
-    #      if all jobs are a success, return success
-    status_job = monitored_jobs 
-             |> Enum.reduce_while(%{:status => "initial", :job => :nojob}, fn(job, acc) ->
-               Logger.info fn ->
-                 "monitored job: #{inspect(job)}"
-               end
-                cond do
-                  job && (job.status =~ "failed" || job.status =~ "canceled") ->
-                    acc = %{status: "failed", job: job}
-                    {:halt, acc}
-                  job && (job.status =~ "running" || job.status =~ "created") ->
-                    # can only go to a running status from initial, running, or success status
-                    if (acc.status =~ "running" || acc.status =~ "initial" || acc.status =~ "success") do
-                      acc = %{status: "running" , job: job}
-                    end
-                    {:cont, acc}
-                  job && job.status =~ "success" ->
-                    # The Backend Dashboard will NOT set the badge status to success when a 
-                    # child -- it's ignored for a child 
-                    # can only go to a success status from initial, running, or success status
-                    if (child_pipeline == false && (acc.status =~ "success" || acc.status =~ "initial")) do
-                      acc = %{status: "success", job: job} 
-                    end
-                    {:cont, acc}
-                  true ->
-                 Logger.error fn ->
-                   "unhandled job status: #{inspect(job)}  not handled"
-                 end
-                    {:cont, acc}
-                end 
-             end) 
-  end 
-
-
-  def badge_url(monitor_job_list, child_pipeline, internal_pipeline_id) do
-     project = Repo.all(from projects in CncfDashboardApi.Projects, 
-                                          left_join: pipelines in assoc(projects, :pipelines),     
-                                          where: pipelines.id == ^internal_pipeline_id) 
-                                          |> List.first
-    Logger.info fn ->
-      "deploy url project: #{inspect(project)}"
-    end
-
-    status_job = monitored_jobs(monitor_job_list, internal_pipeline_id)
-                  |> status_job(child_pipeline)
-
-    Logger.info fn ->
-      "status_job: #{inspect(status_job)}"
-    end
-    Logger.info fn ->
-      "status_job.job: #{inspect(status_job.job)}"
-    end
-
-    if status_job.job != :nojob do
-      Logger.info fn ->
-        "status_job.job != :nojob"
-      end
-      source_key_pipeline_jobs = Repo.all(from skpj in CncfDashboardApi.SourceKeyPipelineJobs, where: skpj.new_id == ^status_job.job.id) |> List.first
-      if source_key_pipeline_jobs do
-        "#{project.web_url}/-/jobs/#{source_key_pipeline_jobs.source_id}"
-      else
-        ""
-      end
-    end
-  end
-
- @doc """
  Gets all project and pipeline information based on the internal
  `project_id` and  `pipeline_id`.
 
@@ -537,11 +389,11 @@ defmodule CncfDashboardApi.GitlabMonitor do
     rm_record = CncfDashboardApi.GitlabMonitor.Dashboard.upsert_ref_monitor(pipeline_monitor, target_pm, target_pl, pipeline_order)
 
 
-    job_names = monitored_job_list("project")
+    job_names = CncfDashboardApi.GitlabMonitor.Job.monitored_job_list("project")
     dbs_record = CncfDashboardApi.GitlabMonitor.Dashboard.update_badge(rm_record,
                                                                        target_pl.ref,
-                                                                       badge_status_by_pipeline_id(job_names, false, "", target_pm.pipeline_id),
-                                                                       badge_url(job_names, false, target_pm.pipeline_id),
+                                                                       CncfDashboardApi.GitlabMonitor.Job.badge_status_by_pipeline_id(job_names, false, "", target_pm.pipeline_id),
+                                                                       CncfDashboardApi.GitlabMonitor.Job.badge_url(job_names, false, target_pm.pipeline_id),
                                                                        1)
 
     # if pipeline_type = "build" then the project_id is a target project
@@ -598,9 +450,9 @@ defmodule CncfDashboardApi.GitlabMonitor do
       end
 
       if cross_cloud_pipeline_monitor do
-        job_names = monitored_job_list("cross-cloud")
-        cc_status = badge_status_by_pipeline_id(job_names, cross_cloud_pipeline_monitor.child_pipeline, cloud.cloud_name, cross_cloud_pipeline_monitor.pipeline_id)
-        cc_deploy_url = badge_url(job_names, cross_cloud_pipeline_monitor.child_pipeline, cross_cloud_pipeline_monitor.pipeline_id)
+        job_names = CncfDashboardApi.GitlabMonitor.Job.monitored_job_list("cross-cloud")
+        cc_status = CncfDashboardApi.GitlabMonitor.Job.badge_status_by_pipeline_id(job_names, cross_cloud_pipeline_monitor.child_pipeline, cloud.cloud_name, cross_cloud_pipeline_monitor.pipeline_id)
+        cc_deploy_url = CncfDashboardApi.GitlabMonitor.Job.badge_url(job_names, cross_cloud_pipeline_monitor.child_pipeline, cross_cloud_pipeline_monitor.pipeline_id)
       end
 
       Logger.info fn ->
@@ -608,9 +460,9 @@ defmodule CncfDashboardApi.GitlabMonitor do
       end
 
       if cross_project_pipeline_monitor do
-        job_names = monitored_job_list("cross-project")
-        cp_status = badge_status_by_pipeline_id(job_names, cross_project_pipeline_monitor.child_pipeline, cloud.cloud_name, cross_project_pipeline_monitor.pipeline_id)
-        cp_deploy_url = badge_url(job_names, cross_project_pipeline_monitor.child_pipeline, cross_project_pipeline_monitor.pipeline_id)
+        job_names = CncfDashboardApi.GitlabMonitor.Job.monitored_job_list("cross-project")
+        cp_status = CncfDashboardApi.GitlabMonitor.Job.badge_status_by_pipeline_id(job_names, cross_project_pipeline_monitor.child_pipeline, cloud.cloud_name, cross_project_pipeline_monitor.pipeline_id)
+        cp_deploy_url = CncfDashboardApi.GitlabMonitor.Job.badge_url(job_names, cross_project_pipeline_monitor.child_pipeline, cross_project_pipeline_monitor.pipeline_id)
       end
 
       cond do
